+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 ## [Unreleased]
 
 ### Added
+ - Add the oihana/core/normalize function : Normalizes a value according to the given cleaning flags.
  - **JSON:**
    - Add the getJsonType function 
  - **Strings:**